18/08

+++ user_router.py
@@ -17,14 +17,17 @@ async def get_User(id:int,
                       db: SessionLocal = Depends(get_db)):
     query = db.query(User).filter(User.id == id).first()
     if query is None:
-        return "Comment not found"
+        raise HTTPException(status_code=404, detail="Comment not found")
     else:
         return query
 
 
 @router.post("/new-User", response_model=user_read)
 async def add_User(schema: UserSchema,
-                      db: SessionLocal = Depends(get_db)):
+                   db: SessionLocal = Depends(get_db)):
+    query = db.query(User).filter(User.name == schema.name).first()
+    if query is not None:
+        raise HTTPException(status_code=404, detail="User already exist")
     model = User()
     model.name = schema.name
     model.password = schema.password
@@ -37,13 +40,12 @@ async def add_User(schema: UserSchema,
 @router.put("/edit-User", response_model=user_read)
 async def edit_User(id:int,
                     schema: UserSchema,
-                    db: SessionLocal = Depends(get_db),
-                    ):
+                    db: SessionLocal = Depends(get_db),):
 
     model = db.query(User).filter(User.id == id).first()
 
     if model is None:
-        raise HTTPException(status_code=404, detail="Not Found")
+        raise HTTPException(status_code=404, detail="User not Found")
     else:
         model_ = User()
         model_.id = id
@@ -59,7 +61,8 @@ async def edit_User(id:int,
 async def del_User(id: int,
                       db: SessionLocal = Depends(get_db)):
     comment = db.query(User).filter(User.id == id).first()
-
+    if comment is None:
+        raise HTTPException(status_code=404, detail="User not found")
     db.delete(comment)
     db.commit()
     return "Successfully deleted"
